fix: :truck: Renamed enemy -> enemies/fighter
Issue: #4
