+++ src/ingame/enemies/fighter/bullet.rs
